+++ Backend/ProjectSem2/src/main/java/fptAptech/theSun/dto/ProductViewDto.java
@@ -19,6 +19,6 @@ public class ProductViewDto implements Serializable {
 
     private Double price;
 
-    private Float discount;
+    private Double discount;
 
 }
